Split web server into modules
