+++ sys/txs-webserver/app/Main.hs
@@ -1,6 +1,6 @@
 module Main where
 
-import Lib
+import API
 
 main :: IO ()
 main = startApp
